Added azure-pipline.yml for CI and PackNuget.ps1 to pack nuget packages.
Removed the Apps.Browser project from the solution as wasn't building,
it didn't seem to have much in it of use anyway. Prefixed all
PackageIds with Blockcore and reset the package versions to 1.0.0.
PackNuget takes a mandatory param 'releaseType' which is appended to
the version with the build number from the environment, the pipeline.

Fix #3